+++ src/WeUIDialog/WeUIDialog.tsx
@@ -10,7 +10,6 @@ import { Fade } from '@material-ui/core';
 interface Props {
   show: boolean;
   title: string;
-  content: string;
 
   leftText?: string;
   onLeftClick?: () => void;
@@ -21,7 +20,7 @@ interface Props {
 const WeUIDialog: FC<Props> = ({
   show,
   title,
-  content,
+  children,
   leftText = '',
   onLeftClick = () => {},
   rightText = '确定',
@@ -35,7 +34,7 @@ const WeUIDialog: FC<Props> = ({
           <div className="weui-dialog__hd">
             <strong className="weui-dialog__title">{title}</strong>
           </div>
-          <div className="weui-dialog__bd">{content}</div>
+          <div className="weui-dialog__bd">{children}</div>
           <div className="weui-dialog__ft">
             <a className="weui-dialog__btn weui-dialog__btn_default" onClick={onLeftClick}>
               {leftText}